python-code/requirements.txt
numpy==1.24.3
librosa==0.9.2
soundfile>=0.10.2,<0.11
pydub==0.25.1
openai-whisper==20230918
googletrans==3.1.0a0
torch>=1.10.0,<1.11.0
torchvision>=0.11.0,<0.12.0
torchaudio>=0.10.0,<0.11.0
python-dotenv==1.0.0
tqdm==4.65.0
pyannote.audio
sounddevice==0.4.6
